requirement.txt
torch 
numpy 
scikit-learn 
matplotlib 
seaborn
prometheus_client 
requests
python-dotenv
psutil